+++ CHANGELOG.md
@@ -43,6 +43,9 @@
   * Added config to change how many scrap stacks are unscrapped per player
   * Added item blacklist config
 
+* Benthic Transform Random Item:
+  * Added config to change how many items are transformed per effect activation
+
 **1.10.1 Changes:**
 
 <details>